Completed Functionality
Got the buttons to work with the display and completed the arithmetic functions so that the expressions can be evaluated. However, the calculator completes the calculations from right to left and consequently does not follow the arithmetic guidelines of PEMDAS.  I also enveloped the code in an IIFE to avoid polluting the global namespace.